+++ tests/helpers/templates/wheel/standard_default_build_script_artifacts.py
@@ -21,7 +21,6 @@ def get_files(**kwargs):
         files.append(f)
 
     files.append(File(Path(kwargs['package_name'], 'lib.so'), ''))
-    files.append(File(Path(metadata_directory, 'entry_points.txt'), ''))
     files.append(
         File(
             Path(metadata_directory, 'WHEEL'),