+++ tests/helpers/templates/wheel/standard_default_python_constraint.py
@@ -20,7 +20,6 @@ def get_files(**kwargs):
 
         files.append(f)
 
-    files.append(File(Path(metadata_directory, 'entry_points.txt'), ''))
     files.append(
         File(
             Path(metadata_directory, 'WHEEL'),